+++ aframe/src/location-based/gps-entity-place.js
@@ -1,17 +1,17 @@
 AFRAME.registerComponent('gps-entity-place', {
     _cameraGps: null,
     schema: {
-        latitude: {
-            type: 'number',
-            default: 0,
-        },
         longitude: {
             type: 'number',
             default: 0,
         },
+        latitude: {
+            type: 'number',
+            default: 0,
+        }
     },
     init: function () {
-        window.addEventListener('gps-camera-origin-coord-set', function() {
+        window.addEventListener('gps-camera-origin-coord-set', function () {
             if (!this._cameraGps) {
                 var camera = document.querySelector('[gps-camera]');
                 if (!camera.components['gps-camera']) {
@@ -22,6 +22,23 @@ AFRAME.registerComponent('gps-entity-place', {
             }
 
             this._updatePosition();
+        }.bind(this));
+
+        window.addEventListener('gps-camera-update-position', function (ev) {
+            if (!this.data || !this._cameraGps) {
+                return;
+            }
+
+            var dstCoords = {
+                longitude: this.data.longitude,
+                latitude: this.data.latitude,
+            };
+
+            var distance = this._cameraGps.computeDistanceMeters(ev.detail.position, dstCoords, true);
+
+            this.el.setAttribute('distance', distance);
+            this.el.setAttribute('distanceMsg', formatDistance(distance));
+            this.el.dispatchEvent(new CustomEvent('gps-entity-place-update-positon', { detail: { distance: distance } }));
         }.bind(this));
 
         this._positionXDebug = 0;
@@ -42,7 +59,7 @@ AFRAME.registerComponent('gps-entity-place', {
      * @returns {void}
      */
     _updatePosition: function () {
-        var position = { x: 0, y: 0, z: 0 }
+        var position = { x: 0, y: this.el.getAttribute('position').y || 0, z: 0 }
 
         // update position.x
         var dstCoords = {
@@ -52,6 +69,7 @@ AFRAME.registerComponent('gps-entity-place', {
 
         position.x = this._cameraGps.computeDistanceMeters(this._cameraGps.originCoords, dstCoords, true);
         this._positionXDebug = position.x;
+
         position.x *= this.data.longitude > this._cameraGps.originCoords.longitude ? 1 : -1;
 
         // update position.z
@@ -62,7 +80,10 @@ AFRAME.registerComponent('gps-entity-place', {
 
         position.z = this._cameraGps.computeDistanceMeters(this._cameraGps.originCoords, dstCoords, true);
         position.z *= this.data.latitude > this._cameraGps.originCoords.latitude ? -1 : 1;
-
+        if(position.y !== 0) {
+            position.y = position.y - this._cameraGps.originCoords.altitude;
+        }
+        
         // update element's position in 3D world
         this.el.setAttribute('position', position);
     },