+++ three.js/src/threex/threex-artoolkitsource.js
@@ -105,7 +105,7 @@ ARjs.Source.prototype._initSourceImage = function (onReady) {
     domElement.style.width = this.parameters.displayWidth + 'px';
     domElement.style.height = this.parameters.displayHeight + 'px';
 
-    onReady();
+    domElement.onload = onReady;
     return domElement
 }
 
@@ -138,7 +138,7 @@ ARjs.Source.prototype._initSourceVideo = function (onReady) {
     domElement.style.width = this.parameters.displayWidth + 'px';
     domElement.style.height = this.parameters.displayHeight + 'px';
 
-    onReady();
+    domElement.onloadeddata = onReady;
     return domElement
 }
 
@@ -182,20 +182,20 @@ ARjs.Source.prototype._initSourceWebcam = function (onReady, onError) {
     navigator.mediaDevices.enumerateDevices().then(function (devices) {
         var userMediaConstraints = {
             audio: false,
-            video: {
-                facingMode: 'environment',
-                width: {
-                    ideal: _this.parameters.sourceWidth,
-                    // min: 1024,
-                    // max: 1920
-                },
-                height: {
-                    ideal: _this.parameters.sourceHeight,
-                    // min: 776,
-                    // max: 1080
-                }
-            }
+            video: true
         };
+
+        if (window.innerWidth < 800) {
+            var width = (window.innerWidth < window.innerHeight) ? 480 : 640;
+
+            userMediaConstraints = {
+                audio: false,
+                video: {
+                    facingMode: 'environment',
+                    width: { min: width, max: width }
+                },
+            };
+        }
 
         if (null !== _this.parameters.deviceId) {
             userMediaConstraints.video.deviceId = {